feature: 폴더 구조 변화에 따른 수정 사항 반영
