+++ src/pages/EditInfoPage.tsx
@@ -1,11 +1,12 @@
 import { useLocation } from 'react-router-dom';
 
 import DesignerEditInfoSection from '@/views/EditInfoPage/components/DesignerEditInfoSection';
+import ModelEditInfoSection from '@/views/EditInfoPage/components/ModelEditInfoSection';
 
 const EditInfoPage = () => {
   const isModel = useLocation();
 
-  return <>{isModel.state ? <div>가콩 여기야</div> : <DesignerEditInfoSection />}</>;
+  return <>{isModel.state ? <ModelEditInfoSection /> : <DesignerEditInfoSection />}</>;
 };
 
 export default EditInfoPage;
